All roles should be associated with sys_role
This roleattribute is used to allow roles to be associated with sys_obj_type associated types.

This basically means that all roles end up authorized to associate with any file.

This might sound excessive but we are not using these roles to enforce access to file objects.
Instead we'll use them for rbacsep and enforce rbacsep constraints where feasible.