fix a typo and make these macros available
one may want to use these for kdevtmpfs, but it is not ideal and i try not to use them.

udevd will take care of labeling any blk/chr nodes kdevtmpfs creates.
allowing kdevtmpfs to create and setattr of generic devtmpfs.fs chr/blk files generally makes the system far more reliable but it is perfect.